+++ frontend/voting/CategoryGroupHeader.vue
