+++ frontend/voting/CategoryGroupTabBar.vue
